+++ src/app/start-coffee-table-group/thanks/page.tsx
@@ -1,0 +1,24 @@
+import DefaultLayout from '@/components/layouts/DefaultLayout';
+import { createMetaData } from '@/util/createMetaData.server';
+
+export const meta = createMetaData({
+	title: 'Volunteer Application Received!',
+	description: `Thank you so much for your willingness to help at VC!`,
+	Hero: 'UndrawPowerful',
+});
+
+export default function Thanks() {
+	return (
+		<DefaultLayout
+			simple
+			Hero="UndrawPowerful"
+			heroHeader={meta.title as string}
+			heroSubheader={meta.description as string}
+		>
+			<div className="lead mb-5">
+				<h2>Thank you so much for your willingness to help at VC!</h2>
+				<p>One of our team members will get back to you shortly.</p>
+			</div>
+		</DefaultLayout>
+	);
+}
